requirements.txt
# requirements.txt (Updated for Auth + DB)

fastapi
uvicorn[standard]
transformers>=4.40.0
torch --index-url https://download.pytorch.org/whl/cpu # Keep CPU torch for deployment size
sentencepiece
accelerate
python-dotenv # For managing secrets like DB connection string, JWT secret

# --- Added for Auth & DB ---
pymongo>=4.0 # MongoDB driver (use pymongo[srv] if needed for DNS seedlist connection)
passlib[bcrypt] # For password hashing
python-jose[cryptography] # For JWT handling
email-validator # Often needed by FastAPI for email validation in Pydantic